requirements.txt
google-cloud-bigquery==3.25.0
requests==2.28.1